# api-java/app/src/main/java/com/jalau/api/domain/dto/LoginRequestDTO.java
package com.jalau.api.domain.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequestDTO {
    @NotBlank
    private String login;
    @NotBlank
    private String password;
}