# api-java/app/src/main/java/com/jalau/api/domain/dto/TokenResponseDTO.java
package com.jalau.api.domain.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class TokenResponseDTO {
    private String token;
}